Re #5 Add support to route packets to connectors based on policy/DB entries
A route can be configured per agent, the route in the database will have the key of
username:route. For example user foobar@nextensio.com wants to influence routing to
www.google.com, the admin will configure a route with a tag saying say "version1"

So the route lookup will be done with foobar@nextensio.com:www.google.com as the key
and the version1 tag can be used to influence which connector is chosen.

To see how route is configured, go to http://server.nextensio.net/, click login,
and click on "edit tenant" on one of the tenants and see the "Route" tab on the
left hand side where we can configure routes.

The connector choosing part is pending