Re #20 OPA helping test scripts with required data
NOTE: The whole thing done here should somehow be under a test environment variable check or maybe we dont even
need to do that, we just log the values into a seperate logger regardless of whether its test or production, I
need some more time to figure this out, opened a ticket https://gitlab.com/nextensio/clustermgr/-/issues/11
to track that seperately

For example, the controller is configured to change some policy, that is the test script is configuring the controller.
Now, the test script cant really test the changes till it knows for sure that controller has received the mongo changeset
notification AND updated its data structures with the new policy. The test script can wait for some arbitrary period of
time, which is what a lot of ppl do in a lot of companies, but those things are fragile and based on timing conditions
they end up behaving erratically, either we wait too long or we wait too less. So to prevent all that, we need OPA to
help the scripts with some information that "ok, I got the update and my databases are ready", so the scripts
then wait for a precise event rather than a random period of time